requirements.txt
# Hogwarts-bench Dependencies

# LLM API client
openai>=1.0.0

# Tokenization
tiktoken>=0.5.0

# Interactive visualizations
plotly>=5.0.0

# Environment configuration
python-dotenv>=1.0.0

# Async HTTP requests
aiohttp>=3.9.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
